--- a/src/geohash_helper.c
+++ b/src/geohash_helper.c
@@ -226,13 +226,11 @@ double geohashGetDistance(double lon1d, double lat1d, double lon2d, double lat2d
     u = sin((lat2r - lat1r) / 2);
     v = sin((lon2r - lon1r) / 2);
     double a = u * u;
-    double sqrt_a = u;
     // if v == 0 we can avoid doing expensive math
     if (v != 0.0){
         a += cos(lat1r) * cos(lat2r) * v * v;
-        sqrt_a = sqrt(a);
     }
-    return 2.0 * EARTH_RADIUS_IN_METERS * asin(sqrt_a);
+    return 2.0 * EARTH_RADIUS_IN_METERS * asin(sqrt(a));
 }
 
 int geohashGetDistanceIfInRadius(double x1, double y1,