--- a/src/geohash_helper.c
+++ b/src/geohash_helper.c
@@ -258,8 +258,11 @@ int geohashGetDistanceIfInRadiusWGS84(double x1, double y1, double x2,
 int geohashGetDistanceIfInRectangle(double width_m, double height_m, double x1, double y1,
                                     double x2, double y2, double *distance) {
     double lon_distance = geohashGetDistance(x2, y2, x1, y2);
+    if (lon_distance > width_m/2) {
+        return 0;
+    }
     double lat_distance = geohashGetDistance(x2, y2, x2, y1);
-    if (lon_distance > width_m/2 || lat_distance > height_m/2) {
+    if (lat_distance > height_m/2) {
         return 0;
     }
     *distance = geohashGetDistance(x1, y1, x2, y2);