If we run the following benchmark
```
memtier_benchmark -c 1 -t 1 --pipeline 1 --test-time 60 --command "GEOSEARCH key FROMLONLAT 7 55 BYBOX 200 200 KM" --hide-histogram
```
and profile it:
![image](https://user-images.githubusercontent.com/5832149/203671764-2d956f47-f350-4dd2-b44d-c11cc6bfb02f.png)

We see that 54.78% of cpu cycles are from geohashGetDistanceIfInRectangle.
Within it we're calling 3x geohashGetDistance. The first 2 times we call them to produce intermediate results.
This PR focus on optimizing for those 2 intermediate results.

- 1 Reduce expensive computation on intermediate geohashGetDistance with same long ( 48895ee2815164749b04a3454c41220c41c6e87c )
- 2 Avoid expensive lat_distance calculation if lon_distance fails beforehand ( 82d33e23b19c9adff97190e37a24ad03376796e8 )

## Results

On pipeline 1, single client benchmark, we move from average latency (including RTT) of 93.59895 ms to 79.07571 ms ( approximately 15% latency drop ).

Furthermore we can see that the command latency distribution is now more stable ( check avg, p50, p99 and p999 for last result )

### baseline from unstable branch ( 3b462ce566e577ffcb35822a0a2372f691326cd4 )
```
ALL STATS
=====================================================================================================
Type            Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
-----------------------------------------------------------------------------------------------------
Geosearchs        10.67        93.59895        97.27900        97.79100        97.79100      1596.95 
Totals            10.67        93.59895        97.27900        97.79100        97.79100      1596.95 
```

### After 1 Reduce expensive computation on intermediate geohashGetDistance with same long ( 48895ee2815164749b04a3454c41220c41c6e87c )
```
ALL STATS
=====================================================================================================
Type            Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
-----------------------------------------------------------------------------------------------------
Geosearchs        11.43        87.40721        87.55100        90.62300        91.13500      1709.95 
Totals            11.43        87.40721        87.55100        90.62300        91.13500      1709.95 
```

### After 2 Avoid expensive lat_distance calculation if lon_distance fails beforehand ( 82d33e23b19c9adff97190e37a24ad03376796e8 )
```
ALL STATS
=====================================================================================================
Type            Ops/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
-----------------------------------------------------------------------------------------------------
Geosearchs        12.66        79.07571        79.35900        79.35900        79.87100      1893.85 
Totals            12.66        79.07571        79.35900        79.35900        79.87100      1893.85 
```