@oranagra 2022-11-24
@filipecosta90 if it simple, i'm curious to know the impact of the reordering we did and update the numbers in the top comment