@filipecosta90 2022-11-28
> @filipecosta90 if it simple, i'm curious to know the impact of the reordering we did and update the numbers in the top comment

@oranagra I've updated the PR comment with the data.
Reorerding moved avg. latency from  79.87100 ms to   74.23900 ms  =) Indeed a good improvement!!